Configure for building CoffeeScript
Uses coffee-script npm module's CLI to generate index.js from
index.coffee. We use npm's scripts as a simple task runner to build.

We also specify the "files" array in package.json to indicate what
should actually be shipped. In this case, we only need index.js and
pepjs.js to be shipped.